Rename dds data reader file
