+++ dds/src/implementation/dds_impl/mod.rs
@@ -29,5 +29,5 @@ pub mod reader_factory;
 pub mod status_condition_impl;
 pub mod status_listener;
 pub mod topic_impl;
-pub mod user_defined_data_reader;
+pub mod dds_data_reader;
 pub mod user_defined_subscriber;
